feat: sending info about new unit
created service for sending data, validator for unit name and showing modal

+++ src/app/features/unit/unit-form/unit-form.component.ts
@@ -10,11 +10,13 @@ import { FormControl, NonNullableFormBuilder, Validators } from '@angular/forms'
 import { MatFormField } from '@angular/material/form-field';
 import { StorageService } from '@shared/services/';
 import { Subscription } from 'rxjs';
+import { UnitFormService } from './unit-form.service';
+import { unitNameTakenValidator } from './unit-name-taken.validator';
 
 @Component({
   selector: 'app-unit-form',
-  templateUrl: 'unit-form.html',
-  styleUrls: ['unit-form.scss'],
+  templateUrl: 'unit-form.component.html',
+  styleUrls: ['unit-form.component.scss'],
   changeDetection: ChangeDetectionStrategy.OnPush,
 })
 export class UnitFormComponent implements OnInit, OnDestroy {
@@ -24,6 +26,7 @@ export class UnitFormComponent implements OnInit, OnDestroy {
 
   private builder = inject(NonNullableFormBuilder);
   private storageService = inject(StorageService);
+  private unitFormService = inject(UnitFormService);
   private subscriptions = new Subscription();
   unitForm = this.createUnitForm();
 
@@ -45,15 +48,23 @@ export class UnitFormComponent implements OnInit, OnDestroy {
 
   onSubmit(): void {
     this.unitForm.markAllAsTouched();
-    ///todo in next task: validate if unitName exist, send form and userId to backend, navigate to confirmation page
-    /// link to next task https://github.com/akademia-envelo-3/MELO-front/issues/51
-    sessionStorage.clear();
+    if (this.unitForm.valid) {
+      sessionStorage.clear();
+      this.unitFormService.sendNewUnitData(
+        this.unitForm.value.unitName!,
+        this.unitForm.value.unitDescription!
+      );
+    } else {
+      this.unitFormService.showErrorSnackBar('Nieprawidłowo wypełniony formualarz');
+    }
   }
 
   private createUnitForm() {
     return this.builder.group({
       unitName: this.builder.control('', {
         validators: [Validators.required, Validators.maxLength(255)],
+        asyncValidators: [unitNameTakenValidator(this.unitFormService)],
+        updateOn: 'blur',
       }),
       unitDescription: this.builder.control('', {
         validators: [Validators.required, Validators.maxLength(4000)],
